+++ src/styles/register/child.ts
@@ -1,0 +1,91 @@
+import { StyleSheet } from 'react-native';
+import colors from '../../constants/colors';
+import fonts from '../../constants/fonts';
+import layout from '../../constants/layout';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: colors.text.white,
+    paddingTop: 60,
+  },
+  scroll: {
+    flex: 1,
+  },
+  scrollContent: {
+    paddingHorizontal: layout.screenPaddingHorizontal,
+    paddingTop: layout.screenPaddingTop,
+  },
+  titleSection: {
+    marginBottom: 24,
+    gap: 8,
+  },
+  title: {
+    fontSize: 26,
+    fontFamily: fonts.bold,
+    color: colors.text.primary,
+    lineHeight: 34,
+  },
+  subtitle: {
+    fontSize: 14,
+    fontFamily: fonts.medium,
+    color: colors.text.secondary,
+    lineHeight: 22,
+  },
+  // 안내 배너
+  banner: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: colors.secondary[500],
+    borderRadius: 15,
+    padding: 15,
+    marginBottom: 14,
+    gap: 8,
+  },
+  bannerIcon: {},
+  bannerText: {
+    flex: 1,
+    fontSize: 12,
+    fontFamily: fonts.medium,
+    color: colors.text.primary,
+    lineHeight: 20,
+  },
+  // 자녀 추가 버튼
+  addButton: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    borderWidth: 1.5,
+    borderStyle: 'dashed',
+    borderColor: colors.primary[300],
+    borderRadius: 15,
+    paddingVertical: 15,
+    gap: 6,
+    marginBottom: 8,
+  },
+  addIconCircle: {
+    width: 16,
+    height: 16,
+    borderRadius: 8,
+    backgroundColor: colors.primary[300],
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  addButtonText: {
+    fontSize: 12,
+    fontFamily: fonts.semiBold,
+    color: colors.primary[400],
+  },
+  scrollBottom: {
+    height: 24,
+  },
+  // 하단 고정 버튼
+  footer: {
+    paddingHorizontal: layout.screenPaddingHorizontal,
+    paddingBottom: layout.screenPaddingBottom,
+    paddingTop: 12,
+    backgroundColor: colors.text.white,
+  },
+});
+
+export default styles;
