+++ src/types/child.ts
@@ -1,0 +1,10 @@
+import { SchoolResult } from './school';
+
+export interface ChildInfo {
+  id: string;
+  name: string;
+  selectedSchool: SchoolResult | null;
+  schoolQuery: string;
+  grade: number | null;
+  calendarColor: string | null;
+}
