+++ src/types/school.ts
@@ -1,0 +1,5 @@
+export interface SchoolResult {
+  name: string;
+  address: string;
+  type: string;
+}
